Rename test folder from AddAppConfigKeyValues... to GacelaConfigAddAppConfig...
